feat: show recent past events
Displaying a list of the 20 most recent past events
